# swagger.yml
openapi: 3.0.0
info:
  title: School Management System
  description: This is a School Management System which provides automation of various academic and administrative operations.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://sch-management-server.herokuapp.com/api
    description: Production Server
paths:
...
  /admin/teacher:
    post:
      tags:
        - Admin
      summary: Create Teacher
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/TeacherSignup'
      responses:
        '200':
          description: Teacher Created Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /admin/fetchTeacher/{teacherId}:
    get:
      tags:
        - Admin
      summary: Fetch Teacher
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teacherId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Teacher fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Teacher Not Found
        '500':
          description: Internal Server Error
  /admin/teacher/{teacherId}:
    delete:
      tags:
        - Admin
      summary: Delete Teacher
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teacherId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Teacher Deleted Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Teacher Not Found
        '500':
          description: Internal Server Error
  /admin/class:
    post:
      tags:
        - Admin
      summary: Create Class
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/Class'
      responses:
        '200':
          description: Class Created Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /admin/fetchClass/{classId}:
    get:
      tags:
        - Admin
      summary: Fetch Class
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Class fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Class Not Found
        '500':
          description: Internal Server Error
  /admin/updateClass/{classId}:
    put:
      tags:
        - Admin
      summary: Update Class
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/Class'
      responses:
        '200':
          description: Class Updated Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Class Not Found
        '500':
          description: Internal Server Error
  /admin/class/{classId}:
    delete:
      tags:
        - Admin
      summary: Delete Class
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Class deleted Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Class Not Found
        '500':
          description: Internal Server Error
  /admin/teacher/{teacherId}/class/{classId}:
    post:
      tags:
        - Admin
      summary: Assign Teacher to Class
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teacherId
          required: true
          schema:
            type: string
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Teacher assigned to a class Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Class/Teacher Not Found
        '500':
          description: Internal Server Error
  /admin/class/{classId}/teachers:
    get:
      tags:
        - Admin
      summary: Fetch Teachers assigned to a particular class
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Teacher fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Class Not Found
        '500':
          description: Internal Server Error
  /admin/student:
    post:
      tags:
        - Admin
      summary: Create Student
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/StudentSignup'
      responses:
        '200':
          description: Student Created Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '500':
          description: Internal Server Error
  /admin/fetchStudent/{studentId}:
    get:
      tags:
        - Admin
      summary: Fetch Student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student fetched Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Student Not Found
        '500':
          description: Internal Server Error
  /admin/student/{studentId}:
    delete:
      tags:
        - Admin
      summary: Delete Student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student deleted Successfully
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Student Not Found
        '500':
          description: Internal Server Error
  /admin/class/{classId}/student/{studentId}:
    post:
      tags:
        - Admin
      summary: Assign a class to a student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: string
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Class assigned to a student Successfully
        '400':
          description: Invalid input
        '401':
          description: Not Authorized to perform this action
        '404':
          description: Class/Student Not Found
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    TeacherSignupInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        isAdmin:
          type: boolean
          default: false
        isTeacher:
          type: boolean
          default: true
        isLibrarian:
          type: boolean
          default: false
    StudentSignupInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        isAdmin:
          type: boolean
          default: false
...
    ClassInput:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: number
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    TeacherSignup:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TeacherSignupInput'
        description: Teacher Signup
...
    StudentSignup:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StudentSignupInput'
        description: Student Signup
...
    Class:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClassInput'
        description: Class Creation
...